added attribute to supertype
